third_party/blink/web_tests/svg/dynamic-updates/SVGFEFloodElement-dom-flood-color-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'flood-color' attribute of the SVGFEFloodElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS floodElement.getAttribute('flood-color') is "rgb(145, 32, 103)"
PASS floodElement.getAttribute('flood-color') is "rgb(64, 128, 103)"
PASS successfullyParsed is true

TEST COMPLETE
